requirements.txt
aiohttp
asyncio
sqlalchemy
datetime
more-itertools
asyncpg
python-dotenv
